@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Demo::Application.config.secret_token = 'e7fab1a73ddd62bfa0a88dc58d8f9e1479d326cc1ec61f877cf448c34ae101e3ef81a2f658c2fa280ddfeadf5bc02520fb090578f829f7477f497fbc916be7b3'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Demo::Application.config.secret_key_base = 'de298625ae543a4d2fa8f868e8c2e58df215cacf43027e6395fa5ae1e8d22dea5772a52eb7152af6bccfbb9bf39b514009eecce277ea204a2c46cfb0e53dbd52'
